+++ core/request_option.go
@@ -68,7 +68,7 @@ func (r *RequestOptions) cloneHeader() http.Header {
 	headers := r.HTTPHeader.Clone()
 	headers.Set("X-Fern-Language", "Go")
 	headers.Set("X-Fern-SDK-Name", "github.com/KardFinancial/kard-go-sdk")
-	headers.Set("X-Fern-SDK-Version", "v13.2.1")
+	headers.Set("X-Fern-SDK-Version", "v13.3.0")
 	headers.Set("User-Agent", "github.com/KardFinancial/kard-go-sdk/0.0.0-fern-placeholder")
 	return headers
 }